Add --helo option to allow overriding the proxy's hostname.
The default behaviour has the proxy forwarding whatever it receives
verbatim. Postfix does not like this:

host [filter] replied to HELO/EHLO with my own hostname mail.example.org
mail for [filter] loops back to myself

By specifying --helo filter.example.org the proxy will change how it
announces itself as a server when a client connects (the 220 response)
and also how it announces itself as a client (HELO / EHLO) when connecting
to the proxy destination.